Created a smaller network (Baby version) to verfiy network works as intended.
My GPU (GTX 1060 3GB) is too small to store the net parameters and input as float so i settle with half floats. Half floats lacked the precision the network needed to converge so I must turn to cloud GPUs for help. I isoloated the half floats as the issue by painfully running the baby network on my cpu and waiting for it to converge or produce favorable results relative to the gpu version.